requirements.txt
# Core Dependencies
xrpl-py==4.4.0              # XRP Ledger Python library
fastapi==0.128.0            # Web framework
uvicorn==0.40.0             # ASGI server
pydantic==2.12.5            # Data validation
python-dotenv==1.2.1        # Environment variable management

# Database
asyncpg==0.31.0             # Async PostgreSQL driver

# Optional (for development)
# pytest==7.4.3             # Testing framework
# pytest-asyncio==0.21.1    # Async test support
# black==23.12.0            # Code formatter
# ruff==0.1.9               # Linter
